swapped minimax to alphaBeta, remade logic between sending pieces to second board

+++ src/com/chess/engine/player/ai/BoardEvaluator.java
@@ -3,5 +3,5 @@ package com.chess.engine.player.ai;
 import com.chess.engine.board.Board;
 
 public interface BoardEvaluator {
-    double evaluate(Board board, int depth);
+    int evaluate(Board board, int depth);
 }
